Added Pattern Recognition
Created a dictionary to keep track of all the values that have appeared. From there I created separate functions that could check the appearance of each number without needing to iterate over the dice multiple times.

Still have to iterate over the dictionary multiple times but it's still fewer.

With this the Challenge is complete! With a little less than 24 hours to spare.